vcxproj rely relation int64 exp/gold NFSOCK

+++ NFMidWare/NFGuildPlugin/NFCGuildComopnent.cpp
@@ -37,7 +37,7 @@ bool NFCGuildComponent::CheckPower(const NFGUID& self, const NFGUID& xGuildID, i
 
 int NFCGuildComponent::OnCreateGuildProcess(const NFGUID & self, const int from, const int event, std::string & arg)
 {
-	CLIENT_MSG_PROCESS_NO_LOG(event, arg.c_str(), arg.length(), NFMsg::ReqAckCreateGuild)
+	CLIENT_MSG_PROCESS_NO_LOG(event, arg.c_str(), (uint32_t)arg.length(), NFMsg::ReqAckCreateGuild)
 
 	std::string strRoleName;
 	int nLevel = 0;
